# spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: Ashurbanipal Protocol
  version: "1"
  description: >
    Read-only SQL table-browsing API (Postgres, MySQL, or SQLite
    backend). Paths are relative to
    {mount}/api, where {mount} is the implementation-defined mount point
    (reference default: /__ashurbanipal). The UI itself is served at
    {mount} as text/html and is not described here.
servers:
  - url: "{mount}/api"
    variables:
      mount:
        default: /__ashurbanipal
        description: Implementation-defined mount point.
paths:
  /sources:
    get:
      operationId: listSources
      summary: List source names selectable as the `source` param elsewhere.
      responses:
        "200":
          description: >
            Live allow-list for the `source` param on /schemas, /tables,
            /table-counts, /tables/data, /tables/common-values, and
            /tables/referenced-by (spec/protocol.md §5.8). Exactly one
            entry for a single-source deployment.
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcesResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /schemas:
    get:
      operationId: listSchemas
      summary: List schema names selectable as the `schema` param elsewhere.
      parameters:
        - $ref: "#/components/parameters/Source"
      responses:
        "200":
          description: >
            Live allow-list for the `schema` param on /tables,
            /table-counts, /tables/data, /tables/common-values, and
            /tables/referenced-by (spec/protocol.md §5.7).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemasResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /tables:
    get:
      operationId: listTables
      summary: List base tables of the resolved schema.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
      responses:
        "200":
          description: Table list (also the allow-list for `table` params).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                type: object
                required: [tables]
                properties:
                  tables:
                    type: array
                    items:
                      $ref: "#/components/schemas/TableInfo"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /table-counts:
    get:
      operationId: tableCounts
      summary: Approximate row counts from pg_class.reltuples.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
      responses:
        "200":
          description: Approximate counts for every table.
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountsResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /tables/data:
    get:
      operationId: tableData
      summary: Paginated, filtered, sorted rows for a single table.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
        - name: table
          in: query
          required: true
          description: >
            Must exactly match (case-sensitive) a table name from /tables;
            otherwise 400.
          schema:
            type: string
          examples:
            # Three named cases, sourced from conformance/seed/seed.sql, so
            # the Examples phase actually exercises distinct 200-response
            # shapes rather than only ever hitting `orders`: a plain table,
            # one with a composite PK/FK pair (composite-key omission from
            # column metadata, spec/protocol.md §5.4.1), and one deliberately
            # left un-ANALYZEd (total_approx == -1, spec/protocol.md §5.4.4).
            orders:
              value: orders
            inventory_counts:
              value: inventory_counts
            feature_flags:
              value: feature_flags
        - name: filter
          in: query
          required: false
          description: >
            URL-encoded JSON array of condition objects (the filter AST —
            never DSL text; DSL parsing is frontend-only, see
            spec/filter-dsl.md). `logic` is absent on the first element and
            required on every subsequent one; `value` is present iff `op`
            is not IS NULL / IS NOT NULL; at most 10 conditions; the
            encoded size is bounded (see spec/protocol.md §5.4.2). An
            empty array means no filter. Structural violations and unknown
            columns yield 400.
          content:
            application/json:
              schema:
                type: array
                maxItems: 10
                items:
                  $ref: "#/components/schemas/FilterCondition"
              examples:
                # Keyed to match the `orders` case above so a combined
                # example request stays internally consistent (a real
                # column of that table).
                orders:
                  value:
                    - column: status
                      op: "="
                      value: completed
        - name: limit
          in: query
          required: false
          description: >
            Clamped to [1, max_page_size] (reference default max 100),
            never an error — including negative or out-of-range values.
            Default is default_page_size (reference 50). Deliberately no
            `minimum`/`maximum` on the schema below: those are JSON Schema
            *rejection* constraints, and this field is never rejected for
            being out of range, only clamped — declaring bounds here would
            claim behavior the API doesn't have.
          schema:
            type: integer
          example: 10
        - name: offset
          in: query
          required: false
          description: >
            Clamped to a minimum of 0, never an error — including negative
            or out-of-range values, same treatment as `limit`. No
            `minimum` on the schema below for the same reason as `limit`:
            it's a JSON Schema rejection constraint, and this field is
            never rejected for being out of range.
          schema:
            type: integer
            default: 0
        - name: sort
          in: query
          required: false
          description: >
            Single column name, validated against the table's real
            columns; unknown column yields 400. Ordering uses the
            column's native type ordering.
          schema:
            type: string
          examples:
            orders:
              value: created_at
        - name: order
          in: query
          required: false
          description: Any value other than asc/desc yields 400.
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: One page of rows plus column metadata.
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableDataResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /tables/common-values:
    get:
      operationId: commonValues
      summary: Most-common values for one column, from pg_stats.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
        - name: table
          in: query
          required: true
          schema:
            type: string
          examples:
            # `feature_flags` is deliberately excluded from the seed's
            # ANALYZE pass (conformance/seed/README.md), so this pair
            # exercises the empty-values-not-an-error path
            # (spec/protocol.md §5.5) alongside the ordinary populated one.
            orders:
              value: orders
            feature_flags:
              value: feature_flags
        - name: column
          in: query
          required: true
          schema:
            type: string
          examples:
            orders:
              value: status
            feature_flags:
              value: enabled
      responses:
        "200":
          description: >
            Most-common values, most frequent first. Empty when the column
            has no planner statistics (never an error).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonValuesResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
  /tables/referenced-by:
    get:
      operationId: referencedBy
      summary: Foreign keys elsewhere in the schema that target this table.
      parameters:
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Schema"
        - name: table
          in: query
          required: true
          description: >
            Must exactly match (case-sensitive) a table name from /tables;
            otherwise 400.
          schema:
            type: string
          examples:
            # Three named cases from conformance/seed/seed.sql exercising
            # distinct 200 shapes (spec/protocol.md §5.9): a multi-referrer
            # target (users — two of the entries share a referencing
            # table), a composite-FK target (inventory_locations — one
            # entry whose `columns` has two pairs), and a target nothing
            # references (feature_flags — `referenced_by` is []).
            users:
              value: users
            inventory_locations:
              value: inventory_locations
            feature_flags:
              value: feature_flags
      responses:
        "200":
          description: >
            One entry per foreign-key constraint pointing at `table`;
            `referenced_by` is [] when nothing does (not an error).
          headers:
            x-ashurbanipal-protocol:
              $ref: "#/components/headers/ProtocolVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferencedByResponse"
        "400":
          $ref: "#/components/responses/ClientError"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    Source:
      name: source
      in: query
      required: false
      description: >
        Selects which registered DbSource this operation targets
        (spec/protocol.md §1). Absent resolves to the first-registered
        source. Present must exactly match (case-sensitive) an entry
        from GET /sources; otherwise 400.
      schema:
        type: string
    Schema:
      name: schema
      in: query
      required: false
      description: >
        Selects which schema this operation is scoped to (spec/protocol.md
        §1). Absent resolves to the connection's own default. Present
        must exactly match (case-sensitive) an entry from GET /schemas;
        otherwise 400.
      schema:
        type: string
  headers:
    ProtocolVersion:
      description: Protocol version; "1" for this document.
      required: true
      schema:
        type: string
        const: "1"
  responses:
    ClientError:
      description: >
        Client error — unknown source/schema/table/column, invalid
        filter, invalid parameter, or a table the role can't read.
      headers:
        x-ashurbanipal-protocol:
          $ref: "#/components/headers/ProtocolVersion"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ServerError:
      description: Server error — database failure or query timeout.
      headers:
        x-ashurbanipal-protocol:
          $ref: "#/components/headers/ProtocolVersion"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    SourcesResponse:
      type: object
      required: [sources]
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceEntry"
          description: >
            Selectable sources, in the same order the `source` param's
            absent case resolves against (first entry is the default).
            Exactly one entry for a single-source deployment.
    SchemasResponse:
      type: object
      required: [schemas]
      properties:
        schemas:
          type: array
          items:
            type: string
          description: >
            Selectable schema names, excluding the engine's own
            system/internal namespaces. Exactly one entry on engines with
            no schema concept above a single database.
    TableInfo:
      type: object
      required: [name]
      properties:
        name:
          type: string
        comment:
          type: string
          description: COMMENT ON TABLE text; omitted (never null) when absent.
...
    TableDataResponse:
      type: object
      required: [columns, rows, total_approx]
      properties:
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
        rows:
          type: array
          description: >
            One object per row, keyed by column name. Every value is a
            JSON string (the engine's own text rendering — uuids, numbers,
            timestamps, JSON all serialize as strings) or null for SQL
            NULL; an undecodable value is the sentinel string
            "<undecodable>". Native JSON numbers/booleans are
            non-conformant.
          items:
            type: object
            additionalProperties:
              type: ["string", "null"]
        total_approx:
          type: integer
          format: int64
          description: >
            Whole-table pg_class.reltuples estimate; unaffected by
            filter; may be stale or -1 before first ANALYZE.
    CountsResponse:
      type: object
      required: [counts]
      properties:
        counts:
          type: array
          items:
            type: object
            required: [table, approx_rows]
            properties:
              table:
                type: string
              approx_rows:
                type: integer
                format: int64
                description: May be stale or -1 before first ANALYZE.
    CommonValuesResponse:
      type: object
      required: [values]
      properties:
        values:
          type: array
          description: Most frequent first; empty when no planner statistics.
          items:
            type: object
            required: [value, freq]
            properties:
              value:
                type: string
                description: >
                  Rendered consistently with row serialization (e.g.
                  booleans as "true"/"false").
              freq:
                type: number
                format: float
                description: Fraction of rows, in (0, 1].
...
    FilterCondition:
      type: object
      required: [column, op]
      properties:
        logic:
          type: string
          enum: [AND, OR]
          description: >
            Absent on the first condition, required on every subsequent
            one. AND binds tighter than OR (SQL precedence); no
            grouping/nesting exists.
        not:
          type: boolean
          default: false
          description: Wraps the condition's SQL fragment in NOT (...).
        column:
          type: string
          description: >
            Validated server-side against the table's real columns
            (case-sensitive) before reaching SQL; unknown column yields
            400.
        op:
          type: string
          enum: ["=", "!=", ">", "<", ">=", "<=", "LIKE", "ILIKE", "IS NULL", "IS NOT NULL"]
        value:
          type: string
          description: >
            Required for every op except IS NULL / IS NOT NULL, for which
            it must be absent. Always bound as a query parameter;
            comparison applies to the column cast to text.
...